+++ node/src/index/mod.rs
@@ -7,3 +7,4 @@ pub mod address;
 pub mod filter;
 pub mod outpoint_spend;
 pub mod silent_payments;
+pub mod stint;